various improvements, currently running on arduino mega

+++ mpu6050.h
@@ -4,5 +4,5 @@ void mpu6050_reset();
 void mpu6050_init();
 uint16_t mpu5060_fifo_count();
 void mpu5060_reset_fifo();
-void mpu6050_get_data(float *y, float *p, float *r, int16_t *gy, int16_t *gp, int16_t *gr);
+void mpu6050_get_data(int16_t *y, int16_t *p, int16_t *r, int16_t *gy, int16_t *gp, int16_t *gr);
